reorganized dx folder languageindependent
