+++ src/plone/app/multilingual/dx/form.py
@@ -3,9 +3,13 @@ from Acquisition import aq_base
 from Products.CMFCore.utils import getToolByName
 from interfaces import ILanguageIndependentField
 from plone.app.multilingual.manager import TranslationManager
+from z3c.form.interfaces import DISPLAY_MODE
 from z3c.form.interfaces import IValue
 from z3c.form.interfaces import NO_VALUE
+from z3c.form.validator import StrictSimpleFieldValidator
+from zope.component import getMultiAdapter
 from zope.interface import implementer
+from zope.pagetemplate.interfaces import IPageTemplate
 
 
 def isLanguageIndependent(field):
@@ -16,6 +20,34 @@ def isLanguageIndependent(field):
         return True
     else:
         return False
+
+
+class LanguageIndependentFieldValidator(StrictSimpleFieldValidator):
+    """Override validator so we can ignore language independent fields,
+       these will be automatically filled later on by subscriber.createdEvent
+    """
+    def validate(self, value, force=False):
+        # always pass
+        pass
+
+
+class LanguageIndependentFieldInputTemplate(object):
+    """Override input template for language independent fields with
+       display widget, because values will be automatically filled
+       by later on by subscriber.createdEvent.
+    """
+    def __init__(self, context, request, view, field, widget):
+        self.context = context
+        self.request = request
+        self.view = view
+        self.field = field
+        self.widget = widget
+
+    def __call__(self, widget):
+        template = getMultiAdapter(
+            (self.context, self.request, self.view, self.field, self.widget,),
+            IPageTemplate, name=DISPLAY_MODE)
+        return template(widget)
 
 
 @implementer(IValue)
